Some gardening, Proxy2TargetInspc refactored (while using on embedded platform)
* Serial_HALemC not ready yet, StringPacked_emC not ready yet (test, gardening)

* os_AtomicAcess: compareAndSet is removed, only compareAndSwap,
  defined in compl_adaption.h because platform depending, TODO for Win32,
  only for x64 clarified, but in using sources clarified.
* Warings for visual studio removed and code improved.
* Gardening in compl_adaption.h (while docu)

+++ emC/Base/ExcThreadCxt_emC.c
@@ -116,7 +116,7 @@ MemC getUserBuffer_ThreadContext_emC ( int size, char const* sign, struct Thread
     setUserBuffer_ThreadContext_emC(alloc, _thCxt);
   }
   { MemUnit* endBuffer = PTR_MemC(threadHeap->bufferAlloc, MemUnit) + threadHeap->bufferAlloc.val;  //Addr after buffer.
-    int sizeFree = endBuffer - threadHeap->addrFree;
+    int sizeFree = (int)(endBuffer - threadHeap->addrFree);
     int mask = 0x1;
     int ix = 0;
     if(size ==0){  //special arguments
